Ahora sí funciona el delete y el update

+++ src/routes.php
@@ -1,4 +1,4 @@
-<?php
+ <?php
 
 use Slim\App;
 use Slim\Http\Request;
@@ -63,7 +63,7 @@ return function (App $app) {
 			if(count($result) == 0){
 				$json = array(
 					"status" => "error",
-					"messaje" => "No hay datos en la base de datos"
+					"messaje" => "No hay información en la base de datos"
 				);
 			}else{
 				$json = array(
@@ -79,19 +79,19 @@ return function (App $app) {
 	});
 
 	$app->post('/update_art', function (Request $request, Response $response, array $args) use ($container) {
+		$param = $request->getParsedBody();
 		$data = array();
 		$json = null;
 
 			try{
-				$data = [
-					"id_producto" => $param['id_producto'],
-					"nombre_producto" => $param['nombre_producto'],
-					"cantidad_existencia" => $param['cantidad_existencia'],
-					"precio" => $param['precio']
-				];
+				
 
 				$sql = "UPDATE productos SET nombre_producto = :nombre_producto, cantidad_existencia = :cantidad_existencia, precio = :precio WHERE id_producto = :id_producto";
 				$stm = $this->db->prepare($sql);
+				$stm->bindParam(':nombre_producto',$param['nombre_producto']);
+				$stm->bindParam(':cantidad_existencia',$param['cantidad_existencia']);
+				$stm->bindParam(':precio',$param['precio']);
+				$stm->bindParam(':id_producto',$param['id_producto']);
 
 				if($stm->execute()){
 					$json = array(
@@ -115,6 +115,8 @@ return function (App $app) {
 		return $this->response->withJson($json,200);
 	});
 
+	/*
+
 	$app->post('/modif_art', function (Request $request, Response $response, array $args) use ($container) {
 		$data = array();
 
@@ -126,29 +128,44 @@ return function (App $app) {
 		return $this->response->withJson($json,200);
 	});
 
+	*/
+
 	$app->post('/delete_art', function (Request $request, Response $response, array $args) use ($container) {
-		$get = $request->getParsedBody();
+		$param = $request->getParsedBody();
+		$json = null;
 
 		try{
-			$sql = "DELETE FROM productos WHERE id_producto = :id_producto";
-			$stm = $this->db->prepare($sql);
-			$stm->bindParam(':id_producto', $param['id_producto']);
+			$stm_select = $this->db->prepare("SELECT * FROM productos WHERE id_producto = :id_producto");
+			$stm_select->bindParam(':id_producto', $param['id_producto']);
+			$stm_select->execute();
+			$result_select = $stm_select->fetchAll();
+			
+			if(count($result_select)==1){
+				$sql = "DELETE FROM productos WHERE id_producto = :id_producto";
+				$stm = $this->db->prepare($sql);
+				$stm->bindParam(':id_producto', $param['id_producto']);
 
-			if($stm->execute()){
-				$json = array(
-					"status" => "success",
-					"message" => "Producto eliminado correctamente"
-				);
+				if($stm->execute()){
+					$json = array(
+						"status" => "success",
+						"message" => "Producto eliminado correctamente"
+					);
+				}else{
+					$json = array(
+						"status" => "error",
+						"message" => "No se pudo eliminar el producto"
+					);
+				}
 			}else{
 				$json = array(
 					"status" => "error",
-					"message" => "No se pudo eliminar el producto"
+					"message" => "El producto especificado no existe"
 				);
 			}
 		}catch(PDOException $e){
 			$json = array(
 				"status" => "error",
-				"message" => "No se pudo eliminar el producto"
+				"message" => $e
 			);
 		}
